cleanup of tests and yet another take on getting macos test passing
